move some files related to the homegrown TCP/IP stack out of httptap.go and into device.go
